+++ scripts/utility/effect.js
@@ -1,5 +1,5 @@
 import {chris} from '../helperFunctions.js';
-export async function preCreateActiveEffect(effect, updates, options, id) {
+export async function itemDC(effect, updates, options, user) {
     if (!updates.changes || !effect.parent || !effect.origin) return;
     if (updates.changes.length === 0) return;
     if (effect.parent.constructor.name != 'Actor5e') return;
@@ -16,4 +16,16 @@ export async function preCreateActiveEffect(effect, updates, options, id) {
     }
     if (!changed) return;
     effect.updateSource({'changes': updates.changes});
+}
+export async function fixOrigin(token, options, user) {
+    if ((game.user.id !== user) || token.actorLink) return;
+    let updates = token.actor.effects.reduce((updates, effect) => {
+        if (!effect.origin) return updates;
+        let origin = fromUuidSync(effect.origin);
+        if (!origin instanceof Item) return updates;
+        let item = token.actor.items.get(origin.id);
+        if (item) updates.push({'_id': effect.id, 'origin': item.uuid});
+        return updates;
+    }, []);
+    await token.actor.updateEmbeddedDocuments('ActiveEffect', updates);
 }